Use recent Xcode version for aarch64 gcc build.
There doesn't seem to be any reason for the recent failure to build
GCC 14.1 for aarch64, but we noticed that the Github macos-14 runner
defaults to an unusual Xcode version (15.0.1).

It would probably make sense to go for slightly less than the latest
version (15.4 at the time of writing), but we can't use 15.3 because
it doesn't provide m4 (it does provide gm4, though).

  * .github/workflows/macos_arm.yml: Set xcode version to 15.1.